requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ase==3.22.1
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.16.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
e3nn==0.5.1
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
huggingface-hub==0.17.3
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
latexcodec==2.0.1
lightning-utilities==0.9.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
# mkl-fft==1.3.1
# mkl-service==2.4.0
monty==2023.9.25
mpmath==1.3.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook_shim==0.2.3
numpy==1.26.0
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
overrides==7.4.0
packaging==23.2
palettable==3.3.3
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
plotly==5.17.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3Dmol==2.0.4
pybtex==0.24.0
pycparser==2.21
Pygments==2.16.1
pymatgen==2023.10.4
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pypaddlepaddle-gpu-lightning==1.8.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.4
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.2
sniffio==1.3.0
soupsieve==2.5
spglib==2.1.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
terminado==0.17.1
threadpoolctl==3.2.0
timm==0.9.7
tinycss2==1.2.1
tomli==2.0.1
paddlepaddle-gpu==1.12.1
paddlepaddle-gpuaudio==0.12.1
paddlepaddle-gpumetrics==1.2.0
paddlepaddle-gpuvision==0.13.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
types-python-dateutil==2.8.19.14
tzdata==2023.3
uncertainties==3.1.7
uri-template==1.3.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
yarl==1.9.2